Implement Gale-Shapley stable matching
This is an adaptation of my much earlier implementation made
for the RosettaCode project. The additional requirements of the
task were moved to the runnableExamples block.

The code was updated, thoroughly checked and commented.

The name for the category is similar to one used for earlier
C# and Lua implementations.